Move csv modifiers into helper
