+++ tests/unit/test_bin_load_sample_csv.py
@@ -19,7 +19,17 @@ def test_data_dir(request):
 @pytest.mark.parametrize("test_data_dir", ["good_csvs"], indirect=True)
 def test_create(connection, dbsession, settings, test_data_dir):
     """Most straightforward load works"""
-    ret = main(connection, settings, ["-d", test_data_dir, "--dev"])
+    ret = main(
+        connection,
+        settings,
+        [
+            "-d",
+            test_data_dir,
+            "--dev",
+            "--duplicates",
+            f"{test_data_dir}/duplicates.txt",
+        ],
+    )
     assert ret == 0
 
     ctct = get_contact_by_email_id(dbsession, "e926425d-0189-4cc2-ac7c-b760659ac62f")
@@ -38,7 +48,17 @@ def test_create(connection, dbsession, settings, test_data_dir):
 @pytest.mark.parametrize("test_data_dir", ["good_csvs_no_amo"], indirect=True)
 def test_create_no_amo(connection, dbsession, settings, test_data_dir):
     """If the amo file is empty, we don't make an amo for this record"""
-    ret = main(connection, settings, ["-d", test_data_dir, "--dev"])
+    ret = main(
+        connection,
+        settings,
+        [
+            "-d",
+            test_data_dir,
+            "--dev",
+            "--duplicates",
+            f"{test_data_dir}/duplicates.txt",
+        ],
+    )
     assert ret == 0
 
     ctct = get_contact_by_email_id(dbsession, "e926425d-0189-4cc2-ac7c-b760659ac62f")
